Рефакторинг кода 1. Подготовка для работы со свечным графиком в CandleChart. 2. Перенос FetchCandleData в Api.
